fixing rates mapping and response type on saving operation

+++ src/main/java/com/au/marlo/demo/api/model/RateValue.java
@@ -1,5 +1,8 @@
 package com.au.marlo.demo.api.model;
 
+import java.math.BigDecimal;
+import java.util.Currency;
+
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.databind.PropertyNamingStrategy;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
@@ -12,6 +15,6 @@ import lombok.Data;
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class RateValue {
 
-    private double australianDolar;
-    private double brazilianReal;
+    private BigDecimal australianDolar;
+    private BigDecimal brazilianReal;
 }
